Conditionally support VLA for MSVC Compilation
MSVC does not support variable length arrays (VLAs). VLA support is
conditionally checked by the preprocessor, and MEOW_FFT_NO_VLA is
defined if it is not supported.

Changes were made where applicable in the header. If supported, VLAs
will be used. If MSVC, _malloca will be used for safer allocation within
the stack frame. Otherwise, alloca will be used for allocation on the
stack.

Note that unlike _malloca, alloca will not fall back on heap allocation
and unsafe usage can result in a stack overflow condition.